AGENTE DE TESTEOS/requirements.txt
pytest==7.4.0
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
colorama==0.4.6
pandas==2.1.0
